requirements.txt
anndata
cappa
cellpose
dask[complete]
dask-cuda==25.2.0
dask-image
dask-jobqueue
geopandas
h5py
numpy
pandas
pillow
roifile
scikit-image
scipy
shapely
tifffile
zarr<3
